requirements.txt
pytesseract==0.3.13
python-multipart==0.0.20
uvicorn==0.34.3
pdf2image==1.17.0
pillow==11.2.1
fastapi==0.115.14
